utils: allow using system wide toolchains cache
Out of the box, when building bootloaders, the scripts detect whether we
have the needed toolchains and download them if needed, next to the
source code.

This is great for development purposes because it only download the
toolchains once.

However, on continous integration builds, where new sources are always
downloaded, we also keep downloading the (non-changing) toolchains.

Introduce an environment variable SYSTEM_WIDE_TOOLCHAINS that points
towards an existing, out-of-tree toolchain location.

This can be used to cache the toolchains instead of always
re-downloading them.

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>