utils: config_value: do not exit when shyaml return 1
config_value function is used widely in all scripts to get values from
entries in a yaml config file.
In some cases the yaml entry is optional and thus the entry is not
present in the file.

By default shyaml will return 1 if the entry is not found.
In all scripts if a command returns a non-zero status we exit from
the program (set -e).
That means if the yaml entry is not found, we exit from the program
which is not what we want.

config_value function has been modified to never exit the program even
if shyaml return non-zero status.

Signed-off-by: Julien Masson <jmasson@baylibre.com>