commit-binaries: Move to common tools
commit-binaries is a generic script to commit some binary blobs (like kernel
modules) into a git repo.

It generates a nice commit message with some metadata information about the
source used to generate these binaries.

Move this script to the common tools [1] so that it can be used by other
projects.

[1] https://gitlab.baylibre.com/baylibre/ci/concourse/pipelines/common
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>