build_optee: Fix OP-TEE platform
OP-TEE platform in the past was hard coded as k3 to be generic, however
in recent years, we mention specific SoC we are attempting to boot on.
This helps enable specific features such as utilizing the h/w RNG in the
platform.

However, we cannot use a common platform name as the platforms are
reused (e.x. am62x is used for am62p etc).

Hence introduce a optee.plat field to describe the specific platform
that the OP-TEE build for the platform uses.

Signed-off-by: Nishanth Menon <nm@ti.com>